feat: Add dbo_merge and dbo_extract scripts
This commit introduces `bin/dbo_merge` and `bin/dbo_extract` scripts to parse and modify Digital Buildings Ontology (DBO) `building_config.yaml` files and UDMI `site_model`s.

The `bin/dbo_merge.py` script merges a DBO file into an existing `site_model`, mapping DBO concepts (GUIDs, types, cloud device ids, connections, translations, links) to UDMI `metadata.json` properties.
The `bin/dbo_extract` wrapper renames the old `tests/dbo/basic/dbo_converter.py` to `bin/dbo_extract.py`, which continues to extract a DBO config from an existing `site_model`.

Test scripts (`tests/dbo/basic/run_test.sh`) have been updated to test both the merging and extraction flows against a temporary output directory, which has been added to `.gitignore`.
