+++ tests/dbo/basic/run_test.sh
@@ -9,7 +9,13 @@ if [ -f "$ROOT_DIR/venv/bin/activate" ]; then
     source "$ROOT_DIR/venv/bin/activate"
 fi
 
-python3 $DIR/dbo_converter.py $DIR/site_model $DIR/extracted_building_config.yaml
+# Test merge: building_config.yaml -> site_model
+TEST_SITE_MODEL="$DIR/test_site_model"
+rm -rf "$TEST_SITE_MODEL"
+$ROOT_DIR/bin/dbo_merge $DIR/building_config.yaml "$TEST_SITE_MODEL"
+
+# Test extract: site_model -> extracted_building_config.yaml
+$ROOT_DIR/bin/dbo_extract "$TEST_SITE_MODEL" $DIR/extracted_building_config.yaml
 
 echo "Original building_config.yaml:"
 cat $DIR/building_config.yaml
@@ -23,5 +29,5 @@ if diff -b -w $DIR/building_config.yaml $DIR/extracted_building_config.yaml; the
     echo "SUCCESS: The extracted building_config.yaml matches the original perfectly!"
 else
     echo "FAILURE: The extracted building_config.yaml does not match the original!"
-    exit 1
+    false
 fi
